requirements.txt
llama_index==0.9.8.post1
numpy==1.26.2
pandas==2.1.3
SQLAlchemy==2.0.23
torch==2.1.1
transformers==4.35.2
setuptools
openpyxl
bitsandbytes==0.41.2
scipy
streamlit
yfinance
langchain
python-dotenv
sentence_transformers